move checks from `storage` and `memdb` to `query`
