+++ packages/core/src/index.ts
@@ -20,4 +20,5 @@ export * from './hierarchy'
 export * from './memdb'
 export * from './client'
 export * from './operator'
+export * from './query'
 export { default } from './component'
